Move operations into individual files in search app
