+++ darts/tests.py
@@ -3,7 +3,7 @@ import datetime
 from django.test import TestCase
 from django.utils import timezone
 
-from .operations import DocumentImport
+# from .operations import DocumentImport
 
 # class DartsTests(TestCase):
 #     def test_equal(self):
